+++ src/game/common/interval.class.ts
