+++ src/game/game.class.ts
@@ -1,12 +1,12 @@
 import { IntervalTimer } from './common/interval.class';
 import { ITankClass, TankClass } from './tank/tank.class';
-import { DELTA_T } from '../../constants';
+import { DELTA_T } from '../constants';
 import { MissilesClass } from './missiles/missiles.class';
 import { MapClass } from './map/map.class';
 import { ACTIONS } from 'src/interfaces/ws';
-import { IClientAction } from '../gateway/actions/client-actions';
-import { WsController } from '../ws.controller';
-import { IServerAction } from '../gateway/actions/server-actions';
+import { IClientAction } from 'src/ws/actions/client-actions';
+import { WsController } from '../ws/ws.controller';
+import { IServerAction } from 'src/ws/actions/server-actions';
 
 //where key - userId
 export type GameTanksConstructor = ITankClass[];
