+++ src/game/tank/weapon.class.ts
