+++ test/utils/ws-client/ws-client.ts
@@ -4,7 +4,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { ACTIONS, ActionTypes, IWsData } from 'src/interfaces/ws';
 import WebSocket from 'ws';
 import readline from 'readline';
-import { IClientAction } from 'src/ws/gateway/actions/client-actions';
+import { IClientAction } from 'src/ws/actions/client-actions';
 import { scatter, bg, fg } from 'ervy';
 import { JwtService } from '@nestjs/jwt';
 import { ROLES } from 'src/constants';
